+++ __tests__/fixtures/blink-api-contract/jadx/com/immediasemi/blink/test/FixtureApi.java
@@ -20,4 +20,7 @@ public interface FixtureApi {
 
     @POST
     Object upload(@Url String url, @Body FixtureBody body, Continuation<? super FixtureResponse> continuation);
+
+    @HEAD
+    Object root();
 }